Ars report for accountant

+++ DufflinMunder/Transactions/Sale.cs
@@ -9,7 +9,7 @@ namespace DufflinMunder.Transactions
         public string SalesAgent { get; set; }
         public string ClientName { get; set; }
         public int ClientId { get; set; }
-        public int SaleAmount { get; set; }
+        public decimal SaleAmount { get; set; }
         public Cycle Recurring { get; set; }
         public int ContractLength { get; set; }
